refactor: Move moz/l10n/message.py -> moz/l10n/message/data.py
