+++ moz/l10n/message/data.py
