+++ moz/l10n/resource/android/parse.py
@@ -19,7 +19,7 @@ from re import compile
 
 from lxml import etree
 
-from ...message import (
+from ...message.data import (
     CatchallKey,
     Expression,
     FunctionAnnotation,
