+++ moz/l10n/resource/android/serialize.py
@@ -20,7 +20,7 @@ from typing import Dict, cast
 
 from lxml import etree
 
-from ...message import (
+from ...message.data import (
     CatchallKey,
     Expression,
     FunctionAnnotation,
