+++ moz/l10n/resource/dtd/parse.py
@@ -22,7 +22,7 @@ from re import DOTALL, MULTILINE, UNICODE, compile
 from sys import maxsize
 from typing import Any
 
-from moz.l10n.message import Message, PatternMessage
+from moz.l10n.message.data import Message, PatternMessage
 
 from ..data import Comment, Entry, Resource, Section
 from ..format import Format
