+++ moz/l10n/resource/dtd/serialize.py
@@ -18,7 +18,7 @@ from collections.abc import Iterator
 from re import UNICODE, compile
 from typing import Any
 
-from moz.l10n.message import Message, PatternMessage
+from moz.l10n.message.data import Message, PatternMessage
 
 from ..data import Entry, Resource
 from .parse import name, re_comment
