+++ moz/l10n/resource/inc/parse.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 from re import compile
 from typing import Any
 
-from moz.l10n.message import Message, PatternMessage
+from moz.l10n.message.data import Message, PatternMessage
 
 from ..data import Comment, Entry, Resource, Section
 from ..format import Format
