+++ moz/l10n/resource/ini/serialize.py
@@ -18,7 +18,7 @@ from collections.abc import Iterator
 from re import search
 from typing import Any
 
-from moz.l10n.message import Message, PatternMessage
+from moz.l10n.message.data import Message, PatternMessage
 
 from ..data import Entry, Resource
