+++ moz/l10n/resource/plain_json/parse.py
@@ -18,7 +18,7 @@ from collections.abc import Iterator
 from json import loads
 from typing import Any
 
-from moz.l10n.message import Message, PatternMessage
+from moz.l10n.message.data import Message, PatternMessage
 
 from ..data import Entry, Resource, Section
 from ..format import Format
