+++ tests/resource/test_android.py
@@ -18,7 +18,7 @@ from importlib_resources import files
 from textwrap import dedent
 from unittest import SkipTest, TestCase
 
-from moz.l10n.message import (
+from moz.l10n.message.data import (
     CatchallKey,
     Expression,
     FunctionAnnotation,
