+++ tests/resource/test_ini.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 from textwrap import dedent
 from unittest import TestCase
 
-from moz.l10n.message import PatternMessage
+from moz.l10n.message.data import PatternMessage
 from moz.l10n.resource.data import Comment, Entry, Resource, Section
 from moz.l10n.resource.format import Format
 from moz.l10n.resource.ini import ini_parse, ini_serialize
