+++ tests/resource/test_webext.py
@@ -18,7 +18,7 @@ from importlib_resources import files
 from textwrap import dedent
 from unittest import TestCase
 
-from moz.l10n.message import Declaration, Expression, PatternMessage, VariableRef
+from moz.l10n.message.data import Declaration, Expression, PatternMessage, VariableRef
 from moz.l10n.resource.data import Entry, Resource, Section
 from moz.l10n.resource.format import Format
 from moz.l10n.resource.webext import webext_parse, webext_serialize
